Extract common `SingleCellAnalyticsSearch` DAO logic to its own DAO class
