+++ app/src/test/java/uk/ac/ebi/atlas/experimentpage/markergenes/MarkerGeneServiceIT.java
@@ -14,7 +14,7 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.context.web.WebAppConfiguration;
 import uk.ac.ebi.atlas.configuration.TestConfig;
-import uk.ac.ebi.atlas.search.CellTypeSearchDao;
+import uk.ac.ebi.atlas.search.celltype.CellTypeSearchDao;
 import uk.ac.ebi.atlas.testutils.JdbcUtils;
 
 import javax.inject.Inject;
