+++ app/src/test/java/uk/ac/ebi/atlas/search/celltype/CellTypeSearchDaoIT.java
@@ -1,17 +1,23 @@
-package uk.ac.ebi.atlas.search;
+package uk.ac.ebi.atlas.search.celltype;
 
 import com.google.common.collect.ImmutableSet;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.core.io.ClassPathResource;
+import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.context.web.WebAppConfiguration;
 import uk.ac.ebi.atlas.configuration.TestConfig;
 import uk.ac.ebi.atlas.solr.cloud.SolrCloudCollectionProxyFactory;
+import uk.ac.ebi.atlas.testutils.JdbcUtils;
 
 import javax.inject.Inject;
+import javax.sql.DataSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -20,10 +26,36 @@ import static org.assertj.core.api.Assertions.assertThat;
 @ContextConfiguration(classes = TestConfig.class)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class CellTypeSearchDaoIT {
+
+    @Inject
+    private JdbcUtils jdbcUtils;
+
+    @Inject
+    private DataSource dataSource;
+
     @Inject
     private SolrCloudCollectionProxyFactory collectionProxyFactory;
 
     private CellTypeSearchDao subject;
+
+    @BeforeAll
+    void populateDatabaseTables() {
+        var populator = new ResourceDatabasePopulator();
+        populator.addScripts(
+                new ClassPathResource("fixtures/scxa_analytics.sql")
+        );
+
+        populator.execute(dataSource);
+    }
+
+    @AfterAll
+    void cleanDatabaseTables() {
+        var populator = new ResourceDatabasePopulator();
+        populator.addScripts(
+                new ClassPathResource("fixtures/scxa_analytics-delete.sql")
+        );
+        populator.execute(dataSource);
+    }
 
     @BeforeEach
     void setUp() {
